requirements.txt
# Top-level requirements for attack_simulator.py
requests>=2.31.0
python-dotenv>=1.0.0
